TypeScript: Fix custom type name conflicts
In elm and morphir, a custom type can be defined with various kinds.
Unfortunately, Elm allows one of the kinds to have the same name as the
overall Type. We have encountered cases where a custom type has just one
kind, and that kind has the same name as the type itself.

In TypeScript, we're representing each kind as an Interface, and the
original Custom Type as a Union of those interfaces. In the cases
described above, this means an Interface statment like
('Interface foo =...') and a 'union' statement with only one value
('type foo = foo') This final line is pointless, and creates a name
conflict in the Typescript code.

This commit ensures that such pointless union statements will be
ommited, preventing this particular type of name conflict.