TypeScript: Add codec generation support
Generic functions for serializing and deserializing fundamental types
are provided in a shared module in `redistributable/`.

These are bound together by the TypeScript backend to produce decoders
for specific types in the IR.

We now use Map<K,V> to represent Morphir.SDK.Dict type, previously we
used Array<[K,V]>. This is a better API for TypeScript users.

The BackendTests.elm are removed, they are obsoleted by the test suite
added in this branch.