TypeScript: Add TypeScript testcases for types and codecs
This adds dev-dependencies to package.json: Mocha, @types/mocha,
gulp-mocha and ts-node.

These testcases import the package generated in the previous
step, so we no longer need to explicitly compile that in Gulpfile. Any
errors in the generated TypeScript API should trigger a compile
failure when we compile the testcase.

The Types test was previously run manually but now it's automated.

BackendTests.elm is removed as it is no longer needed if we test the
generated code.