Spark: Add support for Morphir SDK Aggregations
i.e. Morphir.SDK.Aggregate provides functions to perform multiple
aggregations on the same set of data simultaneously.

This commit adds support for transpiling into Spark calls to
`Morphir.SDK.Aggregation.groupBy fieldName |> Morphir.SDK.Aggregation.aggregate`
with the following limitations:

* The weightedAverageOf operator isn't supported yet.
* Calling `aggregate` without `groupBy` is not supported.
* The 'key' field (i.e. the name of the field grouped by) can't be
  renamed in the output yet (Spark automatically names it by the
  original name of the column grouped by). Use of the 'key' field is
  parsed, but not implemented.
* Individual aggregations cannot be filtered yet (use of `withFilter`
  is parsed but not implemented).

In addition to the Spark Backend functionality, this commit also includes
documentation for developers of this functionality and automated tests
that compare the output in elm to the output in Spark.